fix(cloud-web): address PR review feedback for #1282
Resolve reviewer feedback and rebase onto latest main:

- CHANGELOG: add cloud_web entry under Unreleased / Added
- gateway: fail closed when token is unset (reject all inbound webhooks)
- New(): require public_url when register_url is set for gateway transport,
  mirror the rule in client-side validation, and add a regression test
- forward perm:* card actions via IsPermissionResponse so the engine drops
  stale permission clicks instead of leaking allow/deny into the agent prompt
- ko.json: add Korean translations for the cloud_web setup/fields keys
- split cloudweb_test.go by transport (ws/poll/gateway + shared test helper)
