+++ web/src/i18n/locales/es.json
@@ -233,7 +233,8 @@
       "urlRequiredLongPoll": "Long poll requiere URL base de API",
       "urlRequiredGateway": "Gateway requiere URL base de API o URL de registro (al menos una)",
       "invalidHttpUrl": "{{field}} debe ser una URL http(s) válida o host[:puerto]",
-      "invalidWsUrl": "La URL de WebSocket debe comenzar con ws:// o wss://"
+      "invalidWsUrl": "La URL de WebSocket debe comenzar con ws:// o wss://",
+      "publicUrlRequiredGateway": "La URL pública del webhook es obligatoria cuando se establece la URL de registro"
     }
   },
   "fields": {
