+++ web/src/i18n/locales/ko.json
@@ -227,7 +227,15 @@
     "manualSetup": "수동 설정",
     "manualHint": "{{platform}}의 경우 config.toml에서 자격 증명을 설정하고 서비스를 재시작하세요.",
     "advancedOptions": "고급 옵션",
-    "unsupportedPlatform": "지원하지 않는 플랫폼 유형입니다: {{type}}"
+    "unsupportedPlatform": "지원하지 않는 플랫폼 유형입니다: {{type}}",
+    "cloudWeb": {
+      "urlRequiredWebsocket": "WebSocket 전송에는 API 기본 URL 또는 WebSocket URL이 필요합니다 (하나 이상)",
+      "urlRequiredLongPoll": "롱 폴 전송에는 API 기본 URL이 필요합니다",
+      "urlRequiredGateway": "게이트웨이 전송에는 API 기본 URL 또는 등록 URL이 필요합니다 (하나 이상)",
+      "invalidHttpUrl": "{{field}}은(는) 유효한 http(s) URL 또는 host[:port] 형식이어야 합니다",
+      "invalidWsUrl": "WebSocket URL은 ws:// 또는 wss://로 시작해야 합니다",
+      "publicUrlRequiredGateway": "등록 URL을 설정한 경우 공개 웹훅 URL이 필요합니다"
+    }
   },
   "fields": {
     "botToken": "봇 토큰",
@@ -257,7 +265,18 @@
     "appSecret": "앱 시크릿",
     "sandboxMode": "샌드박스 모드",
     "channelSecret": "채널 시크릿",
-    "channelToken": "채널 액세스 토큰"
+    "channelToken": "채널 액세스 토큰",
+    "transport": "전송 모드",
+    "longPollTimeout": "롱 폴 타임아웃 (ms)",
+    "listen": "수신 주소",
+    "publicUrl": "공개 웹훅 URL",
+    "publicUrlHint": "게이트웨이가 cc-connect 웹훅에 접근하기 위한 공개 URL",
+    "registerUrl": "등록 URL",
+    "registerUrlHint": "선택 사항: 웹훅 콜백을 등록할 게이트웨이 엔드포인트",
+    "cloudWebWsHint": "기본 URL이 설정된 경우 선택 사항이며, 기본값은 /cloud-web/ws입니다",
+    "cloudWebBaseUrlHint": "롱 폴에는 필수이며, WebSocket의 경우 이 값 또는 아래 WebSocket URL을 사용하세요",
+    "eventsPath": "이벤트 경로 (롱 폴)",
+    "sendPath": "전송 경로 (아웃바운드 API)"
   },
   "chat": {
     "noChats": "프로젝트가 없습니다",
