+++ web/src/i18n/locales/zh-TW.json
@@ -233,7 +233,8 @@
       "urlRequiredLongPoll": "Long poll 模式需填寫 API 基礎位址",
       "urlRequiredGateway": "Gateway 模式需填寫 API 基礎位址或註冊 URL（至少一項）",
       "invalidHttpUrl": "{{field}} 必須是有效的 http(s) URL 或 host[:port]",
-      "invalidWsUrl": "WebSocket 位址必須以 ws:// 或 wss:// 開頭"
+      "invalidWsUrl": "WebSocket 位址必須以 ws:// 或 wss:// 開頭",
+      "publicUrlRequiredGateway": "設定註冊 URL 時必須提供公開 Webhook 位址"
     }
   },
   "fields": {
