+++ web/src/i18n/locales/zh.json
@@ -233,7 +233,8 @@
       "urlRequiredLongPoll": "Long poll 模式需填写 API 基础地址",
       "urlRequiredGateway": "Gateway 模式需填写 API 基础地址或注册 URL（至少一项）",
       "invalidHttpUrl": "{{field}} 必须是有效的 http(s) URL 或 host[:port]",
-      "invalidWsUrl": "WebSocket 地址必须以 ws:// 或 wss:// 开头"
+      "invalidWsUrl": "WebSocket 地址必须以 ws:// 或 wss:// 开头",
+      "publicUrlRequiredGateway": "设置注册 URL 时必须提供公开 Webhook 地址"
     }
   },
   "fields": {
